+++ src/nodes/NGSI/attribute-value/attribute-value.js
@@ -69,6 +69,9 @@ const attrValue = async function (param) {
       return Number(res.status);
     } else {
       this.error(`Error while managing attribute value: ${res.status} ${res.statusText}`);
+      if (res.data && res.data.description) {
+        this.error(`Details: ${res.data.description}`);
+      }
       return null;
     }
   } catch (error) {