+++ src/nodes/NGSI/entity/entity.js
@@ -53,6 +53,9 @@ const httpRequest = async function (param) {
       return Number(res.status);
     } else {
       this.error(`Error while managing entity: ${res.status} ${res.statusText}`);
+      if (res.data && res.data.description) {
+        this.error(`Details: ${res.data.description}`);
+      }
       return null;
     }
   } catch (error) {